chapter_004/src/test/java/ru/job4j/lambda/PhoneDictionaryTest.java
package ru.job4j.lambda;

import org.junit.Test;

import java.util.ArrayList;

import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;

/**
 * Test.
 *
 * @author fourbarman (mailto:maks.java@yandex.ru)
 * @version 1
 * @since 14.04.2020
 */
public class PhoneDictionaryTest {
    /**
     * Test when found by name.
     */
    @Test
    public void whenFindByName() {
        PhoneDictionary phones = new PhoneDictionary();
        phones.add(
                new Person("Petr", "Arsentev", "534872", "Bryansk")
        );
        ArrayList<Person> persons = phones.find("etr");
        assertThat(persons.get(0).getSurname(), is("Arsentev"));
    }

    /**
     * Test when Not found.
     */
    @Test
    public void whenNotFoundByName() {
        PhoneDictionary phones = new PhoneDictionary();
        phones.add(
                new Person("Petr", "Arsentev", "534872", "Bryansk")
        );
        ArrayList<Person> persons = phones.find("zzz");
        assertThat(persons.isEmpty(), is(true));
    }
}
